ocr-service/samples/requirements.txt
Pillow==10.4.0
requests==2.31.0

